Existential concepts can now call instance methods
When the existential is created, a method table is populated with
(opaque) function pointers, the first param of which is `Self.Type*`.
When a method is called on a existential I lookup the fn ptr from the
index in the existential vtable, and cast is as the expected function
type, but with the self param of type `i8*`. I call the function,
passing the opaque instance of self as this arg.