Ref counting fixes
Function parameters are passed in at +1, and functions release all
values before returning. The return value is released unowned, so users
of the call can retain the object. If the function return value isn’t
used, we call dealloc_unowned on it, which deallocates it if the ref
count is 0.