State capturing accessor storage
State capturing (for loop thunks, for example) captures the backing
storage, allowing capturing of RefAccessor backed objects. The backing
storage is stored in the global variable, and loading from it allows
access of the original value’s backing storage. This allows mutations
inside the loop to modify the original storage