Refactoring the VIR builder
We don't need all the build methods — use the instruction’s initialiser
instead.
These functions are updated to take Values not Operands, to reduce the
overhead in creating them. I had to add an OpaqueLValue type to provide
the ‘this rval is a ptr lval trust me’ behaviour we could get from a
rval in a PtrOperand.