Finished inliner work
The inliner/explosion system can now reliably inline simple module &
stdlib functions.

To overcome a few problems it was having I:
- Added the setArgs(args:) method to inst — it allows the Inst to
update its own arguments which are stored as instance variables.
- The explosion requires explicitly inserting the tail value, making it
easier to debug and enforcing the semantics needed for it to work
properly
- Correct handling of functions calling other functions, as well as
ordering of the optimisation passes.
- Fixed behaviour of setting Operand.value, this no longer incorrectly
updates Operand.user
- When adding an inst in Explosion.replaceInst it checks whether the
inst is already in the block. This avoids re-adding the parameters of
an inlined function

TODO: Update all Inst types to implement copyInst (and setArgs for
those that need it)