StructFlattenPass can optimise indirect structs with GEP access
Memory of a struct type which is accessed by load/store/gep
instructions can (in some cases) be optimised to not use the memory,
and forward the struct’s elements. This allows optimising the common
case of a struct’s initialiser (good for performance and will help
constant folding work).

Also:
- More test cases
- Inline return value was holding a reference to inlined user — fixed
- ^ allows us to turn the VIR verifier on!

There are still bugs with optimising the stdlib to fix next…