Closures
- Closure parsing, including lookahead for a param list
- Sema, only where type is specified
- Sema introduces a type variable type & a constraint solver. This
isn’t wired up yet but will be used to infer the type of a closure. It
is needed because the type of a closure expr is used to resolve the
function overload; type checking the closure body with variables then
later resolving to a concrete type allows this. The solver cannot
currently substitute the computed type back into the AST, so closure
inference isn’t working yet.
- VIRGen uses the preexisting closure construct, capturing variables is
not working properly yet though
- Fixed a load of problems in VIRGen which passed function pointers
around incorrectly