Constraint system uses type variables to infer function arg types
If a function argument is a closure, we give the closure TypeVariable
param & return types. The constraint system can now use function
overload resolution to constrain the type variables to concrete types
to infer closure types.

Code like this now works:

```
// func map :: Int (Int -> Int) -> Int = …

let y = map 1 (a) do
    return a + 3
```

Above, the `a+3` expression narrows the type of `a` down to void* or
Int, then the return stmt constrains the param type to be the result of
`a+3` and therefore the same type as `a`. After the body is type
checked the `map` function uses these constrains to resolve that the
closure type can be `Int -> Int`.

TODO:
- Better handling of sameTypeVariable constraints — they’re currently
inconsistently applied
- Properly walk a ClosureExpr’s subtree to rewrite its type (and any
manglings) after the type variable is substituted