Dominator frontier set construction
We construct a dominator frontier set for each node in a basic block;
DF(x) contains the first nodes reachable from a block that the block
does not dominate, on each path leaving the block. This will be used in
the mem2reg pass.

Plus I fixed a few errors in dominator tree construction.