RegisterPromotion pass can handle projections of struct/tuple memory
Uses of struct memory can always be replaced by an extract from the
running value; this change allows the mem2reg pass to do this, so more
cases can get this optimisation.
I also change the algorithm for generating the phi set to more closely
match the paper, and change the aggr-flatten pass to work across blocks.