Indirect storage for existential containers
This is the start of an overhaul of existential layout; here we begin
storing existentials on the heap and change some some behaviour:
- Existentials store the type metadata, and runtime’s
`vist_constructExistential` uses the type metadata to move the memory
to the heap.
- The runtime existential representation uses a tagged pointer with a
flag specifying whether the instance needs deallocating.
- The construct_existential instruction takes a value, not an lvalue,
and when lowered allocates & stores into the memory passed into the
runtime. This is so that memory is unique to this call so we don’t get
instances holding multiple mutable references to that slot which can be
deleted by the runtime at any point.
- The existential member access instruction now projects the member,
and returns a pointer, not an offset, simplifying the IR.
- The VIR `vist_constructExistential` can be passed an isLocal param to
specify that the stack value should be used, this is setting up for
future work.

This sets up for future work:
- existential_construct should have an option to construct locally,
i.e. on the stack, when the existential can be shown to not *escape*
the current scope.
- There should be an optimisation pass that converts such constructs to
be local, and an instruction (existential_promote_buffer?) which moves
a stack-stored value to the heap.
- existential_copy and existential_destroy insts which copy and delete
the buffer (the current implementation leaks :/)