Fixes for block abstraction differences
When lowering VIR blocks to LLVM ones, we can split blocks on-the-fly
when we lower cond_fail insts; this was messing up phi uses. We use a
PhiTrackingOperand to allow us to update the reference to the lowered
phi, and each LLVM block holds a set of PhiTrackingOperands which
reference it, which they must update if they wish to split.

We also fix:
- A VIRGen bug which assumed all member lookup insts were viable
lvalues and could be GEP’d.
- When constructing the DFST (for the dom tree) we didn’t check that a
node was not already added, leading to an infinite loop