Existential accessors backed by lvalues, coercedAccessor constructs them
We require any existential accessor be backed by existential memory,
operations to dealloc/export the buffer work on this shared memory.

existential_construct returns this lvalue, and we export whenever this
memory is used. An always-on optimisation will remove all export_buffer
insts unless the memory escapes; this guarantees correctness as the opt
can run when all exit paths are known.