lValueReference and param conventions
An accessor reference now creates a copy when required, lValueReference
allows access to the memory for mutation.

Param conventions are formalised; this will allow me to re-do the
variable release mechanism. What we want is an analysis pass which
releases/destroy_addr the memory if it doesn’t escape to an @out,
@inout param, or a return.