ManagedValue used for VIRGen
VIRGen phase uses a series of ManagedValue types; Managed<Inst> and
AnyManagedValue associate a ‘cleanup’ operation with a value, and uses
must declare how they want to forward this cleanup operation. It also
provides helper methods to bridge abstraction differences in types, for
example bridge indirection differences (i.e. load from a mutable memory
loc for a func call which needs the val) or bridge VIR types for the
coercions allowed by the type system: existentials are constructed and
generics can be substituted.

VIRGenFunction keeps a stack of these cleanup operations, and
VIRGenScope keeps a dict by name; when a scope exits and wants to
release its variables, the VIRGenFunction type walks back along the
list and builds the IR to destroy it.

When constructing the IR, we must be careful that each operation
preserves or correctly forwards the cleanup. For example projecting an
existential buffer, the existential keeps the destructor, but when
constructing the buffer we must create a unique struct instance and it
forwards its ownership to the existential.

——

VIRGen also gains some new protocols: ValueEmitter & _ValueEmitter, and
LValueEmitter & _LValueEmitter. The non underscore variants are
generic, allowing for cleaner code; the underscored variants are used
to dispatch to the non-underscored impls.