VIRGen uses ManagedValue to emit memory safe IR
Here we update VIRGen to properly use the new tools for VIRGen.

We also make some changes to ManagedValue: the cleanup value is
replaced by a `hasCleanup` bool, and the cleanup handle is lazily
calculated when it is destroyed. This is done so when forwarding
cleanup to another managed value, it just needs to set this bit,
instead of working out a possibly different destruction handle (this is
the case if forwarding changes the indirection level)