Introduced Metatype type and typeof functiton
The stdlib can hook into the runtime to get the metadata for a type.

Note: we change mangling here, all symbols get a _V prefix so a string
can be checked if it is a symbol; next we want to mangle types and then
introduce the demangling logic into the runtime.