Parsing, sema, and VIRGen for dynamic check code
This commit adds support for the `as` cast and `x the Type` dynamic
checks.

The as cast statically coerces an expression to another type, `let a =
1 as Any` constructs an existential for Any. For coercions done at a
variable decl site, there is also an implicit coercion expr.

We also introduce checked casts, so you can do:
```
if a the Int do
    print a
```
Which checks the dynamic type of a, and executes a block if it is that
type.

To implement this, we introduce the cast_break VIR instruction, which
calls into the runtime to see if it is that type, and passes that value
to the success block.
```
  cast_break %a: #*Any as Int, $entry.cast0, $entry.exit=

$entry.cast0(%a: #Int):
  %7 = call @_Vprint_tI (%a: #Int)
  break $entry.exit

$entry.exit:
// continue
```