Runtime dynamic casting support
IRGen can call into the runtime to check to see if an existential holds
a concrete type.

`vist_castExistentialToConcrete` returns whether it conforms, and if it
does it memcpy’s the existential buffer into a stack slot. It br's on
the fn return, and in the success case can load the value.

```
  %5 = alloca %Int
  %6 = bitcast %Int* %5 to i8*
  %7 = call i1 @vist_castExistentialToConcrete(%Existential* %4,
%Metadata* @_gInts, i8* %6)
  br i1 %7, label %entry.cast0, label %entry.exit

entry.cast0:                                      ; preds = %entry
  %8 = load %Int, %Int* %5
  call void @_Vprint_tI(%Int %8)
  br label %entry.exit

entry.exit:                                       ; preds =
%entry.cast0, %entry
```