CheckedCastBreakInst can be lowered for all cases
cast_break can check concept -> concept casts too, using the runtime
vist_castExistentialToConcept function to look through metadata. We
must conservatively emit all conformance metadata for every type when a
concept is used in such a cast, as any type could be passed into this
cast.

We also implement the trivial cases of struct -> struct check, which is
statically true or false, and the struct -> concept cast, which if
statically true is lowered to an existential construct call.