AIR and selection DAG work for custom backend
This is the start of work on an experimental backend to replace LLVM.
This commit introduces a low level language to replace LLVM IR: AIR,
and starts work on the next step of the compiler: the AIR DAG tree
which will be used for instruction selection.

This DAG tree will emit a sequence of MachineInsts, which will undergo
register allocation, instruction reordering, then be lowered for each
target platform to object code. Work is inspired by LLVM, the Dragon
compiler book, and Webkit’s AIR backend.