DAG constructed from AIR; DAG matching for instruction selection
The DAG can now be constructed from AIR, I’m still trying to work out
the representation/abstraction level of air, so there’s a lot of messy
work.

I also test out a DAG matching system, which works when tested on some
simple graphs. For this to select instructions we need a way of the
compiler to dynamically test all possible patterns, and also to rewrite
the tree when it has emitted an instruction.