Properly represent the move constraints & freezes in IRC
This commit fixes the problems with the initial IRC implementation: we
reintroduce 2 move lists per node and keep an active flag.

We also respect the precoloredCoalescingHeuristic function, use the
same alias method as the paper, properly implement the coalesce
function, properly manage the many lists used, correctly include
precoloured registers in the determination of k, and better represent
freezing all a node's moves, or just one.