deinit functions implemented for reference types
Ref types can specify a deinit function to be run when the instance is
destroyed.

```
ref type X {
    var x: Int
    deinit = do print x
}
let x = X 1
```
Logs
```
→alloc  	0x100202ab0, rc=1
→release	0x100202ab0, rc=0x0
→dealloc	0x100202ab0
   ↳destructor_fn=0x100000e10
1
```

This is implemented by storing type metadata in the class instance,
creating a destructor function that calls the user defined deinit, then
telling the runtime to call that when deallocating the instance.

The above examples create the following VIR for the destruction:
```
func @_Vdestroy_rPX.refcounted : &runtime (@inout #*X.refcounted) -> #Builtin.Void {
$entry(%self: #*X.refcounted):
  %0 = call @_Vdeinit_dPX (%self: #*X.refcounted)
  %1 = tuple () 	// user: %2
  return %1 // id: %2
}

func @_Vdeinit_dPX : &deinit (@in #*X.refcounted) -> #Builtin.Void {
$entry(%self: #*X.refcounted):
  %0 = struct_element %self: #*X.refcounted, !object 	// user: %self.raw
  %self.raw = load %0: #**X 	// user: %x
  %x = struct_element %self.raw: #*X, !x 	// user: %1
  %1 = load %x: #*Int 	// user: %2
  %2 = call @_Vprint_tI (%1: #Int)
  %3 = tuple () 	// user: %4
  return %3 // id: %4
}
```
The _Vdestroy_rPX.refcounted function ptr is stored in the metadata of the class
instance, so when releasing the class to refcount=0, the deallocation calls it.