class_project_instance inst and ClassType type
We introduce the class_project_instance to project the instance of the
class ptr, and the ClassType class which is the lowered type of a ref
type.

Together they allow checked operations on ref types, fix some bugs in
reference counting, and will allow ref types to become member types of
other user types.