AllocStackPromoter handles CopyAddrInsts; across module deallocators; ref type string
This commit combines lots of work to allow string to use a ref counted
core type:
- Deallocator calls can be emitted if the type decl isn’t visible, like
when importing stdlib types.
- When coercing types, ManagedValue makes sure it is dealing with a
RefType* type before calling vist_retainObject/vist_releaseObject.
- Methods correctly manage the lifetime of self: it is an @in parameter
so gets released at the end of the scope, and its elements are lvalues
so aren’t managed by the scope.
- AllocStackPromoter can properly optimise copying insts, treating them
as a mutation, and correctly replacing a copy inst that doesn’t mutate
the optimised alloc with a store.