Correct managing lifetimes of mutated variables
When changing the value of a variable, we must immediately run the
teardown of the old val before changing it.
Managed val now keeps track of whether a variable is uninitialised,
here this teardown should not be run.
This fixes some specific bugs which were causing leaks for
copy-on-write strings.